chore: pre-launch cleanup (Tier 1 + Tier 2)
Tier 1 (blocking/high-risk):
- Add missing `name` column on core.calcuttas via migration
- Fix float64 arithmetic in ValidateEntry (use int for integer math)
- Add policy authorization tests (CanInviteToCalcutta, CanCreateEntry,
  CanEditEntryBids, CanViewEntryData) — 27 new tests
- Add toDatetimeLocalValue + useBidding hook tests — 49 new tests
- Remove console.warn from CalcuttaEntriesPage production code
- Add InvestmentModel.params validation with allowed-keys-per-kind check

Tier 2 (structural improvements):
- Add missing deleted_at columns, set_updated_at triggers, fix dangerous
  ON DELETE CASCADE to RESTRICT, rename 11 legacy index prefixes
- Fix N+1 queries in HandleListCalcuttasWithRankings (batch tournament
  fetch) and HandleListTournaments (batch ListWinningTeams + schools)
- Split CalcuttaEntriesPage into BiddingOpenView component
- Extract useEntryOwnershipData hook from EntryTeamsPage
- Move AdminBundlesPage API calls to adminService + types/admin
- Split readers.py into readers_ridge.py + readers_predictions.py
- Add _postprocess_predictions edge case tests (7 new tests)
- Document derived.* table dependencies for Python readers

+++ backend/internal/app/calcutta/entry_validation.go
@@ -39,11 +39,11 @@ func ValidateEntry(calcutta *models.Calcutta, entry *models.CalcuttaEntry, teams
 	}
 
 	// Rule 5: Total bids cannot exceed starting budget
-	totalBids := 0.0
+	totalBids := 0
 	for _, team := range teams {
-		totalBids += float64(team.BidPoints)
+		totalBids += team.BidPoints
 	}
-	if totalBids > float64(calcutta.BudgetPoints) {
+	if totalBids > calcutta.BudgetPoints {
 		return fmt.Errorf("total bids cannot exceed starting budget of %d points", calcutta.BudgetPoints)
 	}
